+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## 2.18.3RC1
 * Exit if accessed directly
-* RESOLVE 62 Bootstrap class name hide conflict
+* RESOLVE #62 Bootstrap class name hide conflict
 
 ## 2.18.2
 * RESOLVE #57 Testimonials Stacking on first load