+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## 1.0.4RC1
 * Register `jquery-style`
+* Update copyright year
 
 ## 1.0.3
 * RESOLVE Fatal error: Call to undefined function add_settings_error() in class-aihrus-settings.php on line 559