+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 * Add option "Exclude bxSlider CSS?"
 * Exit if accessed directly
 * Initially no show bxSlider widget testimonials entries
+* RESOLVE #61 Horizontal transitions not working in widget
 * RESOLVE #62 Bootstrap class name hide conflict
+* RESOLVE #65 Widget blank options aren't saving correctly
 * Update copyright year
 
 ## 2.18.2