+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## master
 
-## 1.0.3RC2
+## 1.0.4RC1
+* Register `jquery-style`
+* RESOLVE michael-cannon/testimonials-widget#65 Widget blank options aren't saving correctly
+* Update copyright year
+
+## 1.0.3
 * RESOLVE Fatal error: Call to undefined function add_settings_error() in class-aihrus-settings.php on line 559
 * RESOLVE Featured image via Gravatar not saving correctly
 * Revise data deletion description - Thank you Mark